separate grouping tests out of datagridtests
